DoNotRespond/requirements.txt
pyasn1==0.4.5
pysmb==1.1.27
scapy==2.4.2
pyfiglet